PS v1.0.19: Fix redundant border + eliminate ES2020 syntax
When the preview tab is collapsed, the bottom border of its header tab
and that of the container element overlapped. This fix addresses this issue.

Also, I was inadvertently using the optional chaining operator `obj?.prop`
that is a feature of ES2020. This is too new and hence I have replaced
all occurrences of it.