TT v1.3.7 Fix `addLinks()` further for specific pages
Special:GlobalContributions doesn't generate toollinks for the user
link below the heading. This means the link shouldn't be picked up
manually on this page.

On some other special pages, it turns out toollinks are generated as
siblings of the user link, while the parent is a <bdi> element. It's
therefore necessary to ensure that the subsequent siblings don't
contain any `.mw-usertoollinks` elements, before switching the
reference node to the parent element.